+++ app/Http/Controllers/UserController.php
@@ -149,7 +149,9 @@ class UserController extends Controller
                     $user->foto_profil = User::$FILE_DESTINATION . '/' . $filename;
                 }
                 else {
-                    unlink($user->foto_profil);
+                    if(file_exists($user->foto_profil)) {
+                        unlink($user->foto_profil);
+                    }
                     $user->foto_profil = User::$FILE_DESTINATION . '/' . $filename;
                 }
